# Add small file read optimization