### What does this PR do?

Add small file read optimization


```diff
    117.997 ( 0.001 ms): Bun Pool 14/2950662 openat(dfd: CWD, filename: 0x586eca22)                                = 8
-   117.999 ( 0.001 ms): Bun Pool 14/2950662 statx(dfd: 8, filename: 0x4d328e, flags: EMPTY_PATH, mask: TYPE|MODE|ATIME|MTIME|CTIME, buffer: 0x753a586eb880) = 0
    118.000 ( 0.001 ms): Bun Pool 14/2950662 pread64(fd: 8, buf: 0x5c706baf900, count: 767)                        = 766
    118.002 ( 0.001 ms): Bun Pool 14/2950662 pread64(fd: 8<path>, buf: 0x5c706bafbfe, count: 1, pos: 766) = 0
    118.003 ( 0.001 ms): Bun Pool 14/2950662 close(fd: 8<path>) = 0
```

### How did you verify your code works?